Bug 1149066 max tasks

+++ workers/docker-worker/test/integration/task_polling_test.js
@@ -6,16 +6,17 @@ import DockerWorker from '../dockerworker';
 import TestWorker from '../testworker';
 
 suite('Task Polling', () => {
+
   // Ensure we don't leave behind our test configurations.
-  teardown(settings.cleanup);
-
-  // Total of number of tasks to run...
-  let PARALLEL_TOTAL = 2;
-  let worker;
-
   teardown(async () => {
-    await worker.terminate();
+    if (worker) {
+      await worker.terminate();
+      worker = null
+    }
+    settings.cleanup();
   });
+
+  let worker;
 
   test('do not poll if diskspace threshold is reached', async () => {
     settings.configure({
@@ -23,8 +24,10 @@ suite('Task Polling', () => {
         diskspaceThreshold: 50 * 1000000000
       }
     })
+
     worker = new TestWorker(DockerWorker);
     await worker.launch();
+
     let taskTpl = {
       payload: {
         image: 'taskcluster/test-ubuntu',
@@ -52,5 +55,46 @@ suite('Task Polling', () => {
     }
 
     assert.ok(!claimedTask, 'Task should not have been claimed');
+  });
+
+  test('poll for tasks beyond Azure queue limit', async () => {
+    const QUEUE_LIMIT = 32;
+    const CAPACITY = QUEUE_LIMIT + 4;
+
+    settings.configure({
+      capacity: CAPACITY,
+      capacityManagement: {
+        diskspaceThreshold: 1
+      }
+    });
+
+    worker = new TestWorker(DockerWorker);
+    worker.setMaxListeners(CAPACITY);
+    await worker.launch();
+
+    let tasks = [];
+
+    for (let i = 0; i < CAPACITY; i++) {
+      tasks.push(worker.postToQueue({
+        payload: {
+          features: {
+            localLiveLog: false
+          },
+          image: 'taskcluster/test-ubuntu',
+          command: cmd(
+            "ls"
+          ),
+          maxRunTime: 60 * 60
+        }
+      }));
+    }
+
+    let results = await Promise.all(tasks);
+
+    assert.equal(results.length, CAPACITY, `all ${CAPACITY} tasks must have completed`);
+    results.forEach((taskRes) => {
+      assert.equal(taskRes.run.state, 'completed');
+      assert.equal(taskRes.run.reasonResolved, 'completed');
+    });
   });
 });
